Changes from deployers

+++ roles/explorer.nix
@@ -1,4 +1,4 @@
-pkgs: variant: { name, config, ... }:
+pkgs: variant: { name, config, options, ... }:
 with pkgs;
 
 let
@@ -37,11 +37,17 @@ in {
   services.cardano-postgres.enable = true;
   services.postgresql = {
     ensureDatabases = [ "cexplorer" ];
+    initialScript = builtins.toFile "enable-pgcrypto.sql" ''
+      \connect template1
+      CREATE EXTENSION IF NOT EXISTS pgcrypto SCHEMA pg_catalog;
+    '';
     ensureUsers = [
       {
         name = "cexplorer";
         ensurePermissions = {
           "DATABASE cexplorer" = "ALL PRIVILEGES";
+          "ALL TABLES IN SCHEMA information_schema" = "SELECT";
+          "ALL TABLES IN SCHEMA pg_catalog" = "SELECT";
         };
       }
     ];
@@ -62,16 +68,19 @@ in {
     hostAddr = "127.0.0.1";
   };
 
-  services.graphql-engine.enable = true;
+  services.graphql-engine = {
+    enable = true;
+  };
   services.cardano-graphql = {
     enable = true;
     inherit cardanoNodeConfigPath;
-    genesisByron = nodeCfg.nodeConfig.ByronGenesisFile;
-    genesisShelley = nodeCfg.nodeConfig.ShelleyGenesisFile;
     allowListPath = cardano-explorer-app-pkgs.allowList;
     metadataServerUri = globals.environmentConfig.metadataUrl or null;
     ogmiosHost = ogmiosCfg.hostAddr;
     ogmiosPort = ogmiosCfg.port;
+  } // lib.optionalAttrs (options.services.cardano-graphql ? genesisByron) {
+    genesisByron = nodeCfg.nodeConfig.ByronGenesisFile;
+    genesisShelley = nodeCfg.nodeConfig.ShelleyGenesisFile;
   };
 
   services.cardano-rosetta-server = {
@@ -94,7 +103,7 @@ in {
     allProducers = if (globals.topology.relayNodes != [])
         then [ globals.relaysNew ]
         else (map (n: n.name) globals.topology.coreNodes);
-    totalMaxHeapSizeMbytes = 0.4375 * config.node.memory * 1024;
+    totalMaxHeapSizeMbytes = 0.25 * config.node.memory * 1024;
   };
 
   services.cardano-db-sync = {
@@ -102,7 +111,7 @@ in {
     cluster = globals.environmentName;
     environment = globals.environmentConfig;
     socketPath = nodeCfg.socketPath;
-    logConfig = iohkNix.cardanoLib.defaultExplorerLogConfig // { hasPrometheus = [ "0.0.0.0" 12698 ]; };
+    logConfig = iohkNix.cardanoLib.defaultExplorerLogConfig // { PrometheusPort = globals.cardanoExplorerPrometheusExporterPort; };
     user = "cexplorer";
     extended = globals.withCardanoDBExtended;
     inherit dbSyncPkgs;
@@ -111,11 +120,13 @@ in {
     };
   };
 
-  users.users.cexplorer.isSystemUser = true;
+  users.users.cexplorer = {
+    isSystemUser = true;
+    # so that it can write socket file:
+    extraGroups = ["cardano-node"];
+  };
 
   systemd.services.cardano-db-sync.serviceConfig = {
-    # Put cardano-db-sync in "cardano-node" group so that it can write socket file:
-    SupplementaryGroups = "cardano-node";
     # FIXME: https://github.com/input-output-hk/cardano-db-sync/issues/102
     Restart = "always";
     RestartSec = "30s";
@@ -123,12 +134,10 @@ in {
 
   systemd.services.cardano-ogmios.serviceConfig = {
     User = "cexplorer";
-    SupplementaryGroups = "cardano-node";
   };
 
   systemd.services.cardano-rosetta-server.serviceConfig = {
     User = "cexplorer";
-    SupplementaryGroups = "cardano-node";
   };
 
   systemd.services.cardano-graphql = {
@@ -138,8 +147,6 @@ in {
     serviceConfig = {
       User = "cexplorer";
       RuntimeDirectory = "cardano-graphql";
-      # Put cardano-graphql in "cardano-node" group so that it can write socket file:
-      SupplementaryGroups = "cardano-node";
     };
   };
 
@@ -147,10 +154,12 @@ in {
     environment = {
       HASURA_GRAPHQL_LOG_LEVEL = "warn";
     };
+    # FIXME: run under cexplorer (remove sudo use)
+    #serviceConfig.User = "cexplorer";
   };
 
   systemd.services.cardano-submit-api.serviceConfig = lib.mkIf globals.withSubmitApi {
-    # Put cardano-db-sync in "cardano-node" group so that it can write socket file:
+    # Put cardano-submit-api in "cardano-node" group so that it can write socket file:
     SupplementaryGroups = "cardano-node";
   };
 
@@ -166,6 +175,15 @@ in {
 
   systemd.services.dump-registered-relays-topology = let
     excludedPools = lib.concatStringsSep ", " (map (hash: "'${hash}'") globals.static.poolsExcludeList);
+    get_latest_synced_block = writeText "extract_relays.sql" ''
+      select array_to_json(array_agg(row_to_json(t))) from (
+        select COALESCE(ipv4, dns_name) as addr, port from (
+          select min(update_id) as update_id, ipv4, dns_name, port from pool_relay inner join pool_update ON pool_update.id = pool_relay.update_id inner join pool_hash ON pool_update.hash_id = pool_hash.id where ${lib.optionalString (globals.static.poolsExcludeList != []) "pool_hash.view not in (${excludedPools}) and "}(
+            (ipv4 is null and dns_name NOT LIKE '% %') or ipv4 !~ '(^0\.)|(^10\.)|(^100\.6[4-9]\.)|(^100\.[7-9]\d\.)|(^100\.1[0-1]\d\.)|(^100\.12[0-7]\.)|(^127\.)|(^169\.254\.)|(^172\.1[6-9]\.)|(^172\.2[0-9]\.)|(^172\.3[0-1]\.)|(^192\.0\.0\.)|(^192\.0\.2\.)|(^192\.88\.99\.)|(^192\.168\.)|(^198\.1[8-9]\.)|(^198\.51\.100\.)|(^203.0\.113\.)|(^22[4-9]\.)|(^23[0-9]\.)|(^24[0-9]\.)|(^25[0-5]\.)')
+            group by ipv4, dns_name, port order by update_id
+        ) t
+      ) t;
+    '';
     extract_relays_sql = writeText "extract_relays.sql" ''
       select array_to_json(array_agg(row_to_json(t))) from (
         select COALESCE(ipv4, dns_name) as addr, port from (
@@ -176,9 +194,10 @@ in {
       ) t;
     '';
     relays_exclude_file = builtins.toFile "relays-exclude.txt" (lib.concatStringsSep "\n" globals.static.relaysExcludeList);
-    networkMagic = (builtins.fromJSON (builtins.readFile globals.environmentConfig.nodeConfig.ShelleyGenesisFile)).networkMagic;
   in {
-    path = [ config.services.postgresql.package jq netcat curl dnsutils ];
+    path = config.environment.systemPackages
+      ++ [ config.services.postgresql.package jq netcat curl dnsutils ];
+    environment = config.environment.variables;
     script = ''
       set -uo pipefail
 
@@ -193,10 +212,10 @@ in {
 
         while IFS= read -r ip; do
           set +e
-          PING="$(timeout 7s ${cardano-ping}/bin/cardano-ping -h $ip -p $port -m ${toString networkMagic} -c 1 -q --json)"
+          PING="$(timeout 7s cardano-ping -h $ip -p $port -m $NETWORK_MAGIC -c 1 -q --json)"
           res=$?
           if [ $res -eq 0 ]; then
-            echo $PING | ${jq}/bin/jq -c > /dev/null 2>&1
+            echo $PING | jq -c > /dev/null 2>&1
             res=$?
           fi
           set -e
@@ -227,46 +246,62 @@ in {
         done <<< "$allAddresses"
       }
 
-      excludeList="$(sort ${relays_exclude_file})"
-      cd $STATE_DIRECTORY
-      rm -f *-relay.json
-      i=0
-      for r in $(psql -t < ${extract_relays_sql} | jq -c '.[]'); do
-        addr=$(echo "$r" | jq -r '.addr')
-        port=$(echo "$r" | jq -r '.port')
-        allAddresses=$addr$'\n'$(dig +nocookie +short $addr A)
-        excludedAddresses=$(comm -12 <(echo "$allAddresses" | sort) <(echo "$excludeList"))
-        nbExcludedAddresses=$(echo $excludedAddresses | wc -w)
-        if [[ $nbExcludedAddresses == 0 ]]; then
-          ((i+=1))
-          pingAddr $i $addr $port &
-          sleep 0.5
-        else
-          >&2 echo "$addr excluded due to dns name or IPs being in exclude list:\n$excludedAddresses"
+      run() {
+        epoch=$(cardano-cli query tip --testnet-magic $NETWORK_MAGIC | jq .epoch)
+        db_sync_epoch=$(psql -t --command="select no from epoch_sync_time order by id desc limit 1;")
+
+        if [ $(( $epoch - $db_sync_epoch )) -gt 1 ]; then
+          >&2 echo "cardano-db-sync has not catch-up with current epoch yet. Skipping."
+          exit 0
         fi
+
+        excludeList="$(sort ${relays_exclude_file})"
+        cd $STATE_DIRECTORY
+        rm -f *-relay.json
+        i=0
+        for r in $(psql -t < ${extract_relays_sql} | jq -c '.[]'); do
+          addr=$(echo "$r" | jq -r '.addr')
+          port=$(echo "$r" | jq -r '.port')
+          allAddresses=$addr$'\n'$(dig +nocookie +short $addr A)
+          excludedAddresses=$(comm -12 <(echo "$allAddresses" | sort) <(echo "$excludeList"))
+          nbExcludedAddresses=$(echo $excludedAddresses | wc -w)
+          if [[ $nbExcludedAddresses == 0 ]]; then
+            ((i+=1))
+            pingAddr $i $addr $port &
+            sleep 0.5
+          else
+            >&2 echo "$addr excluded due to dns name or IPs being in exclude list:\n$excludedAddresses"
+          fi
+        done
+
+        wait
+
+        if test -n "$(find . -maxdepth 1 -name '*-relay.json' -print -quit)"; then
+          echo "Found a total of $(find . -name '*-relay.json' -printf '.' | wc -m) relays to include in topology.json"
+          find . -name '*-relay.json' -printf '%f\t%p\n' | sort -k1 -n | cut -d$'\t' -f2 | tr '\n' '\0' | xargs -r0 cat \
+            | jq -n '. + [inputs]' | jq '{ Producers : . }' > topology.json
+          mkdir -p relays
+          mv topology.json relays/topology.json
+          rm *-relay.json
+        else
+          echo "No relays found!!"
+        fi
+      }
+
+      while true
+      do
+        run
+        sleep 3600
       done
-
-      wait
-
-      if test -n "$(find . -maxdepth 1 -name '*-relay.json' -print -quit)"; then
-        echo "Found a total of $(find . -name '*-relay.json' -printf '.' | wc -m) relays to include in topology.json"
-        find . -name '*-relay.json' -printf '%f\t%p\n' | sort -k1 -n | cut -d$'\t' -f2 | tr '\n' '\0' | xargs -r0 cat \
-          | jq -n '. + [inputs]' | jq '{ Producers : . }' > topology.json
-        mkdir -p relays
-        mv topology.json relays/topology.json
-        rm *-relay.json
-      else
-        echo "No relays found!!"
-      fi
     '';
+    startLimitIntervalSec = 1800;
     serviceConfig = {
       User = cfg.user;
       StateDirectory = "registered-relays-dump";
+      Restart = "always";
+      RestartSec = "30s";
+      StartLimitBurst = 3;
     };
-  };
-  systemd.timers.dump-registered-relays-topology = {
-    timerConfig.OnCalendar = "daily";
-    wantedBy = [ "timers.target" ];
   };
 
   services.nginx = {
@@ -415,9 +450,6 @@ in {
           "/relays" = {
             root = "/var/lib/registered-relays-dump";
           };
-          "/metrics2/exporter" = {
-            proxyPass = "http://127.0.0.1:8080/";
-          };
           "/metrics2/cardano-graphql" = {
             proxyPass = "http://127.0.0.1:3100/metrics";
           };
@@ -435,14 +467,15 @@ in {
     {
       job_name = "explorer-exporter";
       scrape_interval = "10s";
-      metrics_path = "/metrics2/exporter";
-      labels = { alias = "explorer-exporter"; };
+      port = globals.cardanoExplorerPrometheusExporterPort;
+      metrics_path = "/";
+      labels = { alias = "${name}-exporter"; };
     }
+  ] ++ lib.optional config.services.cardano-graphql.enable
     {
       job_name = "cardano-graphql-exporter";
       scrape_interval = "10s";
       metrics_path = "/metrics2/cardano-graphql";
       labels = { alias = "cardano-graphql-exporter"; };
-    }
-  ];
+    };
 }
